Bug#37603579 Valgrind reports leaks in mysqltest when shutdown times out
If we get this during valgrind runs:
   mysqltest: At line 168: Command "shutdown_server" failed with error 2. my_errno=175.
then Valgrind will report "still reachable" for the local variable ds_file_name
in do_shutdown_server(). The function calls die(), which does not return,
and hence the apparent leak.

Silence this warning, so we can continue running the rest of the test suite.

Change-Id: I03dfda493a847b3de5755e0217c9b555a933670c