Bug#37560280 Incorrect behavior with EXPLAIN and subqueries
EXPLAIN could execute a stored function in a subquery when the subquery
created a derived table, e.g.:
  EXPLAIN SELECT * FROM (SELECT f()) AS a;

This issue was caused by Table_ref::materializable_is_const(), which
incorrectly treated a function call as a constant expression.
Now, we ensure that a derived table subquery containing a stored
function call is not considered constant.

Change-Id: Id73d567b3617da1523996e8ee0ed73b4879451ad