Adding service debug REST interface
Debugging in JavaScript will be enabled/disabled at service level
through the following API:

Get State:

    GET <service_url>/_debug
    PUT <service_url>/_debug With request body as {"enabled": <true|false>}

A debug port should be configurable on the router configuration.
Since this is only valid when configuring the router in a development
environment, the port configuration should NOT be stored in the
metadata schema.

A new configuration option has been added into the MRS section:

developer_debug_port=<port>

Which will be used by graal to launch a debug server when the debug
is enabled for a given service.

Change-Id: I90ee4421eb446ca3b3a7b3ef6e91bec5c1f4c22b