Adding debug support for JavaScript files
Debugging requires the configuration of the developer_debug_port
setting in the MRS configuration section of the router configuration
file.

When the port is enabled, debugging for a given service can be
enabled/disabled through the following REST endpoint:

   PUT <service_url>/_debug

The request body should be a json document with a single boolean
"enable" entry, i.e.

   {"enabled": true}

When debug is enabled, to debug a given end point on the service
a configuration profile is required in VSCode as follows:

        {
            "name": "Attach to Graal JS",
            "type": "node",
            "request": "attach",
            "port": <developer_debug_port>,
            "trace": true, // NOTE: Use it to diagnose debugging problems....
            "remoteRoot": "<content set root path for built scripts>",
            "localRoot": "<local path to built scripts>"
        }

When debugging is enabled, a debug server is started, and the client
should get connected to it.

After done, normal debugging can be done by setting the needed breakpoints and
triggering the end point to be debugged.

Change-Id: Ibd9e2c9a2ca6acd019adb83820a47d15c1265c40