Bug#37594545 AT testMgm -n TestConcurrentGracefulStop fails sporadically
When restart times out there is not really a way to safely determine if
restart have been initiated or not, even less information if it will
complete or be aborted or stop half way.

In NdbRestarter::restartNodes function instead of trying to convert a
timeout to failure or success a new return code -2 is introduced to
indicate to calling test function that restart timed out.

Change-Id: I23f48062cd0fff85ca5fe58614a3831ece3c919d