Bug# 37607195 - fprintf_string not using the actual quote parameter (mysql-trunk)
The fprintf_string function present in mysqldump takes the quote
of the string as a parameter, but does not pass it to the
mysql_real_escape_string_quote to escape the string.

The fix is to pass the string quote to mysql_real_escape_string_quote
as a parameter.

Added a test case.

Change-Id: I462c5cf46341f9326a18324e85892600a7c08395