Bug#37617852: Assertion `false && "Inconsistent row counts for different AccessPath objects.
The hypergraph optimizer could generate invalid plans when using
indexes on materialized derived tables, potentially causing assert
failures or wrong results.

The problem occurred when the values to look up in an index could come
from two or more tables. The hypergraph optimizer would propose one
index lookup for both alternatives, but it would mark both of the
subplans as dependent on the same table, even though they in reality
were depending on different tables. This could make the hypergraph
optimizer choose an invalid join order with the actual dependency
table inner to the index lookup, or it could cause inconsistencies
when estimating the cardinality of a subplan.

Fixed by making MaterializedPathCache not return cached access paths
with parameter tables that conflict with the index lookup that is
being considered. This makes the hypergraph optimizer create a fresh
access path with a parameter tables property that reflects the actual
dependencies of the index lookup.

Change-Id: Ia5a6f4e1f749c7da0ce6a5664fece6c3e7d369a9