WL#16484  Make InnoDB container aware
Description:
  This worklog makes InnoDB aware of the limits set by the container
when the server is run in a container. It changes the APIs used by
server to read the number of logical CPUs and available physical memory.
InnoDB uses these functions to determine the appropriate default values
of some of the configurations.
  The changes to the APIs will attempt to read the limits as if it were
running in cgroup v2. If this read fails, then it will attempt to read
it as if it were running in cgroup v1. If this fails too, then we assume
that the server is not running in a container, and we fall back to our
traditional methods of invoking different system APIs.
  The APIs are exposed through component library_mysys.

Changes:
  The following configurations will be affected by this change; when the
server is run either in a container, or it is run with restricted
resources (for example, number of CPUs limited with taskset) depending
on the changes in the number of logical CPUs or available physical
memory:

1. The InnoDB configurations whose default values are calculated based
on the number of logical CPUs:
  1.1. innodb_buffer_pool_instances (and innodb_page_cleaners)
  1.2. innodb_purge_threads
  1.3. innodb_read_io_threads
  1.4. innodb_parallel_read_threads
  1.5. innodb_redo_log_capacity only when innodb_dedicated_server is ON
  1.6. innodb_log_writer_threads

2. The configurations whose default values are calculated based
on the available physical memory:
  2.1. temptable_max_ram
  2.2. innodb_buffer_pool_size only when innodb_dedicated_server is ON

Remarks:
  The case of no limits being set by the container and failure to read
the limits set by the container, are treated as the same - fallback to
existing APIs to read required parameters.

Change-Id: Ibda3b72fa64203b4c5e9886c75a80c85d317dc08