Bug#37635908 Relay Log Sanitizer may cause PITR to become stuck
Problem
-------

It was observed that relay log sanitizer checks transaction
boundaries before start point of sanitization is established
which leads to incorrect setting of the start point of
sanitization and incorrect setting of transaction boundaries.
This may lead to marking a valid relay log file as malformed
and if any of the previous relay log files is considered as
valid, cause the relay log sanitizer to remove the
valid relay log files marked as malformed in case any of
the previous relay log files is considered as valid
by the relay log sanitizer.

Analysis / Root-cause analysis
------------------------------

The relay log sanitizer scans the relay log index file backwards.
Each of the relay log files is then scanned forwards.
Since one transaction may span multiple relay log files,
when starting analysis of one file, the relay log sanitizer does
not know whether its current position is inside or outside
a transaction boundary. Therefore, we establish a starting
point for the sanitization process, where we are certain
that the sanitizer is outside a transaction boundary.
This start point is either the source's relay log rotation
event or the end of a transaction:
- a "COMMIT" query
- a "ROLLBACK" query
- an "XA ROLLBACK..." query
- an "XA COMMIT..." query
- an Xid event
- an atomic DDL query

From this start point, the relay log sanitizer seeks out the
last, possibly truncated transaction. However, a flaw has been
discovered in the sanitizer's logic: it fails to bypass transaction
boundary parsing until the starting point is reached. This oversight
can lead to incorrectly flagging a valid relay log file as malformed,
prompting the sanitizer to delete it. As a result, if a process
awaits the application of the relay log while the receiver thread is
disconnected, it may become stuck due to the removal of the valid
file. Notably, this issue affects only relay log sanitization;
binary log recovery begins at the start of the binary log file
and remains unaffected.

Solution
--------

Fixed issues with transaction boundary parsing in the log
sanitizer.

Change-Id: Icc3754079fa337df4c52a1277eba72e8cdb3e0be