WL#16076 step 1: check.py: New feature to list includes
WL#16076: New GTID set data structure

Introduce a feature in check.py that shows the files included from a
given file. This is useful for debugging. The feature is invoked using
the new --list-file-inc-recursive option.

Example:
```
./libs/scripts/check.sh --list-file-inc-recursive \
  ./libs/mysql/binlog/event/control_events.h
```

This will traverse the DAG of included files and print their names in
*post-order*, i.e., included file before including file. This reflects
the order the files are compiled, assuming the #include directives
precede other code.

Change-Id: I70e282b01d8028cf297d15fe68efbc73fbb747bb