Bug#35308309 : Prepare exec invalid variable assertion error
Description:
------------
Server exits abruptly when we try to set the read
only variable INNODB_PURGE_THREADS.

Analysis:
---------
While setting a read-only variable using the SET
command, mysqld checks if the variable is read
only and then errors out with a proper error
message. This is missing when we set the same read
only variable using PREPARED STATEMENTS. So, the
INNODB_PURGE_THREADS is successfully modified and
it leads to this unexpected behaviour.

Fix:
----
Added the check to error out from setting the read
only variables when it is from PREPARED STATEMENT.

Change-Id: I4a41ab4ae42d6e5867ef383671e9444cb044da8f