BUG#36421684: mysql server 8.3.0 heap-buffer-overflow at Multisource_info::get_mi
When using JSON fuctions as a parameter to a
SOURCE_POS_WAIT/MASTER_POS_WAIT, something that should also apply to
other functions that return text results, the string extracted from
the parameter did not have a safe pointer to it.

The cause is that the length of the string was not properly marked
inside its allocated space.

Usage of the method c_ptr_safe fixes this issue.

Change-Id: Ic2c54999293aa2e0833594754ad681d7453e03a1