BUG#37661257 assertion in jit-executor at execute
This patch improves the exception handling, specially in the boundaries
between polyglot and C++, where some exception types were not being
handled correctly, leading to crashes on specific events.

Also introduces a better handling not only of memory errors but any
critical resource that may lead to malfunction of the context.

Also removes the on demand garbage collection logic which now is
unnecessary since establishing the RAM limit gives automatic GC
as needed.

Additionally updates the default RAM size to be used to 1GB.

Change-Id: Id5c685cf923357d9b9dad9f17cd3374896e030a5