Bug#37656707:rapid.autodb_advisor_analyze_data fails when run              with --hypergraph option
Some of the autodb tests are failing when run with hypergraph.
Note that the failure is not consistently seen. So there is
no additional test case added as part of the fix.
The failing queries in the tests involve performance schema
tables and sorting with row-ids. These failures are not seen
with the old optimizer. The difference between the plan
generated by the old optimizer and hypergraph for these
queries is that the old optimizer materializes the result
before the sort with the row-ids. This makes the sort use
row-ids from the temporary table to do the sorting. So we
think the reason for failure in case of hypergraph could be
that, when row-ids are used to fetch the actual table data
once the sort is done, the data is no more valid.
So the fix is to mark the access path involving performance
schema tables as UNSAFE for sorting with row-ids. This
forces the optimizer to materialize the results for such
cases before the sort.

Change-Id: I37b93022cbe77491f69370d5a12db569dcfa1678