BUG#37672626 crash at startup after config-change in jit_executor
When a max ram usage is defined in the configuration, it will be
determined how many ram per JavaScript context will be assigned.

When starting up, since no contexts have been started, this leads
to a zero division, causing the crash in linux.

This patch fixes the logic so the calculation is done only if there
are active contexts.

Change-Id: Ib1169486f3f76e38a9e987073325912dc9fe2c1b