Bug#37657052: mysqld crash at Field::is_nullable while querying ...
In the hypergraph optimizer, if a hidden item is of the form "@var :=
val", then in the temp table used for materialization the replaced
item didn't inherit the 'hidden' attribute from the original item.
Fixed this.

E.g. if there is a GROUP BY @var := val, the temp table included this
item as a visible field even if it was not SELECTed.  Moreover, an
outer query using this SELECT as a subquery crashed because the
subquery had wrong number of SELECT items.

Change-Id: Ibe2e515ecd8116ad7b9161ec376f9c85668516ec