Bug#37617773 Cross DDL operations on MySQL partitioned tables, parallel              COUNT(*) does not report errors.
Desscription:
-------------
When a transaction is going on in a session and we
do a SELECT COUNT(*) on a partition table whose
table definition is changed in another client
session, it is supposed to fail. But the SELECT
COUNT(*) is executed without any errors.

Analysis:
---------
In case of a normal table, the same scenario will
produce an error because we check if the index is
usable when getting the record count. Incase of a
partition table, this check of index usable or not
is missing.

Fix:
----
Add a check to verify if the index is usable when
we are getting the record count.

Change-Id: Ic52ca9688ecd4efc2450e22e5bbe7f617341166d