WL#15440 MySQL REST Service (MRS) - Umbrella WL
Post-push fix: adding a proper synchronization to db_object/modify_fields.inc

When the fields definitions are modified using this script the intention
is that this is an atomic change. There are several SQL statements there
so they should be wrapped inside a transaction so that MRS did not see
a partial change which leads to a random test failures.

Change-Id: I1313a2607060836c6c342354c5be699b25722973