Bug#35889583: Server fails when we execute set of queries Bug#35996409: Failure in Query_expression::is_set_operation() Bug#36404149: Failing the server
Problem is that we attempt to inspect a query expression object that
has been abandoned due to condition elimination. The problem is with
equality operations that are transformed into semi-join conditions,
where we store the left and right parts of the equality in
sj_inner_exprs and sj_outer_exprs.

The fix is to remove the references in the semi-join expression arrays
when removing the corresponding equality condition. A dive within all
join nests of the query block is necessary, due to the fact that we
may have seen multiple transformations that add join nests so far.

Change-Id: I7ae6aca62564486d0261a65afdf54ba057c9687d