Bug#37690477 Router fails to use server that is higher version than its own version.
Fix for bug "Router does not check if server version is higher or not" was
part of 9.2 release.
It introduced a behavior in the Router where it fails to bootstrap against
or use metadata in the runtime from a server that it is a higher version
than it's own. The idea was that it can only promise a backward compatibility,
not compatibility with the future versions.

It was now decided that this behavior is too strict and should be changed.

This patch changes the behavior so that a warning is issued but the server
is still used for fetching the metadata.

Change-Id: I3c71b1ac0b39d36be8c5157695a5671ebba012bd