BUG#37676790 Scripting REST end point returning stalled results
There was a major issue with stalled results on a released context.
i.e. in the case processing ended with a time-out, even the request
was considered complete, because the context remains alive for the
execution of additional requests, the responses associated to the
first request were still generated and consumed by the next request
handled by the context.

To address this, logic was needed to ensure stalled responses after
a context is released are discardedi before putting back the context
in the pool, this change made evident a second problem: the release
logic is synchronous, so the client did not get any response but until
the context was really returned to the pool.

To address this second problem, the a release queue and a release thread
were added to the pool, once a context is done processing a request, it
is added to the queue, and the pool release thread waits for the context
to be available to return it to the pool.

Change-Id: I8973b202265f862f824dc378b07d546235ad266b