Bug#37539314 - HTTP request and response should be logged conditionally [POSTFIX]
Description
===========
Following test was toggling at PB2: developer_request_log.
The test was restarting router, and waiting for the router to
update its entry in `router` table. The problem is that
router does it twice, before first fetch of metadata and
after successful fetch. Thus we do not have the information
if HTTP endpoints are available after restart.

(POST-PUSH FIX WL#15440)

Change-Id: Ia923c8933c03ff75392eb5d24d69f88ecafc1d9c