BUG#37613510: Ever growing GR Transactions Rows Validating after secondary joins the group
Group Replication start operation checks if there are partial
transactions on the `group_replication_applier` channel from a
previous group participation. If partial transactions are found,
`group_replication_applier` channel is stopped after applying all
complete transactions, its relay logs purged and then the channel is
restarted. After this step, distributed recovery kicks-in and
applies the missing data from a group member.
The Group Replication pipeline operation to stop the
`group_replication_applier` channel was incorrectly stopping the
periodic task from the certifier module, which was causing that some
periodic internal operations were not taking place.
One of the tasks that was not happening was the periodic send of the
committed transactions, which omission was preventing the
certification info garbage collection, which on its turn was causing
the continuous increase of the column
COUNT_TRANSACTIONS_ROWS_VALIDATING of the table
performance_schema.replication_group_member_stats.

To solve the above issue, the pipeline operation to stop the
`group_replication_applier` channel now does not interfere with the
certifier module.

Change-Id: I1a4c2f7a5b6d0ca65caf43eeae38103a17b2d5ec