BUG#37697570 BUG#37697649 Defining RAM usage limits in jit-executor
By defining minimum and maximum allowed RAM limits the following
issues get resolved:

- BUG#37697649 infinite loop at GET mrs script
- BUG#37697570 too large maximumRamUser leads to HTTP 501

The ram to be used by the jit-executor is defined as follows:

- Default RAM: 25% of system memory
- Maximum Allowed: 80% of system memory
- Minimum Allowed: 10 MB

If configured RAM is above the max allowed, the jit-executor will
run with defaults, which means it will use as much as needed with
a limit of 80% of the system memory.

If configured RAM is below the min allowed, the jit-executor will
use the default value of 25% of the system memory.

Change-Id: I69cce62b0e7a39234c3ad59431817b8947b4deda