WL#15440 MySQL REST Service (MRS) - Umbrella WL
Bug#37522336 Please consider using UTF8MB4 in order to be unambiguous

Using UTF8 leads to the following server warning:

Warning (code 3719): 'utf8' is currently an alias for the character
set UTF8MB3, but will be an alias for UTF8MB4 in a future release.
Please consider using UTF8MB4 in order to be unambiguous.

This patch replaces use of utf8 with utf8mb4 to get rid of the
ambiguity warning.

Change-Id: Id396e0fa54973e22e64ca628b4fadaf2368d20b2