WL#16898 - HeatWave Temporary Tables recognizes only ENGINE [patch 1]
This worklog changes the engine semantics of temporary tables:
A temporary table must have as single ENGINE as its handled,
therefore SECONDARY_ENGINE cannot be set for it.

A HeatWave Materialized Temporary Table used to be created
with SECONDARY_ENGINE=RAPID expression. Now, it can be created
with ENGINE=RAPID expression. A table created like this will
have its primary engine set to RAPID.
With this change in place, users can also use
default_tmp_storage_engine as a way to specify materialized
temporary tables in HeatWave.

Main changes:
- Primary engine of a temporary table can be also set to RAPID.
Before this patch, it was still InnoDB.
validate_secondary_engine_temporary_table checks if the create
table statement is for a temporary table and whether
it has a secondary engine as its handlerton. If so, it sets
HA_CREATE_INFO::secondary_engine.

Additional:
- CREATE TEMPORARY TABLE .. SECONDARY_ENGINE=RAPID
syntax has been deprecated in favor of
CREATE TEMPORARY TABLE .. ENGINE=RAPID
because we set RAPID as the primary engine of
the temporary table now.
- Block setting 'default_storage_engine' to a
handlerton, which is a secondary engine.
- default_tmp_storage_engine can be set to a
handlerton, which is a secondary engine.

Change-Id: If37bc7d67fbde47358e6ebc5ce264ff36f9fc815