Bug#36816216: mysqld started without --plugin-dir, but with --basedir not ending in / can't lo
When expanding the plugin dir from the basedir in trying to find the
manifest files at server startup the non-canonicized directory name
for the basedir was used. This resulted in an invalid path while
processing the manifest data if the basedir didn't have a terminating
directory separator.
Fixed by using the canonicized name.

Change-Id: I293f17aaead74ac615f648c64e8fdc4a31d155a9