Bug#37628911 Crash caused by race condition between MVCC::view_open and purge view
Problem:

MVCC::view_open() can reopen a read-view which it used before,
if it seems to be still up to date w.r.t. to trx_sys state - i.e. if no
new transaction committed since it was created. But once it makes
decission to reopen, it is crucial that purge coordinator was already
informed to not purge the undo logs needed to support this read-view.
Therefore MVCC::view_open() first marks the view as open, and only then
makes a decission, perhaps closing it again if it can't be reused.
Alas, this could lead to an opposite problem, in which the purge
coordinator selected such a view which was temporarily marked as open,
even if it was very old, i.e. saw less than the previous puge_sys->view.
This could lead to various problems in places where we traverse undo log
chains, as purge_sys->view "traveling back in time" gave impression that
some undo log records should still be available, even though they were
purged, or - in extreme - the whole undo space was truncated already.

Solution:

The purge coordinator should only switch the purge_sys->view to a new
one, if it has strictly larger low_limit_no(). While it's easy to see
that this is sufficient to prevent "travel back in time", and that it
shouldn't impact speed of purge, it's a bit more challanging to prove
that such restrictive rule still ensures that any read view which is in
use or becomes reused, has no problem accessing necessary undo logs.
This patch adds the proof of correctness.

We would like to thank Yin Peng from Tencent who contributed the bug
report, repro scripts, and the sketch of solution.

Change-Id: I86888c896b499a34dd7dec1cf736bde4094039a7