Bug#37656849: `path->init_cost() >= path->init_once_cost()' in operator at join_optimizer/join_optimizer.cc
An assertion failed in the hypergraph optimizer after planning a
complex join. The assertion failed because a hash join had estimated
the init-once-cost to be slightly higher than the init-cost. Since
init-once-cost represents the portion of init-cost that is only paid
on the first initialization, it should not be higher than init-cost.

The problem was a linear interpolation used in calculating the
init-once-cost in ProposeHashJoin(), in which floating-point
imprecision made it return a value just outside of the range in which
the interpolation was done. Fixed by using std::lerp() instead of
rolling our own code for linear interpolation. std::lerp() gives
better guarantees about the result being within range.

While at it, some implementations of linear interpolation in
cost_model.cc are also replaced by calls to std::lerp().

No test case is added for this bug. The repro is quite big, and it is
difficult to reduce it while still getting the floating-point values
to align in a way to trigger the bug. Also, the cost values are likely
to change as the cost model is refined, so it is unlikely that the
exact same test case will expose the bug in the future.

Change-Id: Id5bfa6a02c2b7e87432c6fffede7cdc560f3141c