Bug#37722667 Cluster/J logging updates
In the Cluster/J Logger, add versions of the info(), trace(), debug()
and detail() methods that take a Supplier<String>, as introduced in
java.util.logging in Java 1.8.

In the plain-string versions of these four methods, in JDK14LoggerImpl,
add a runtime assert that the appropriate level has been enabled.

In I18NHelper, add a supplier() method that returns a Supplier<String>
for the given message.

In the test suite, a new class LogAtWarning can be used to configure the
log level to WARNING as a way to test the new asserts at finer levels.
This can be used with:
  java -Djava.util.logging.config.class=testsuite.clusterj.LogAtWarning
The clusterj test in MTR will use this option, since it is both more
strict and less verbose than before.

The rest of this patch consists of fixups to various logging calls
throughout Cluster/J.

Change-Id: Ia43bd14249e113416c9d0b87c9b0e173aff60e8b