WL#16719 Hypergraph cost model for hash-join with spill-to-disk [1/4]
Change the main.opt_hints_join_order_hypergraph test so that the result
file shows just the join order, not full explain output. That makes it
simpler to verify that the test result is correct even if plans change.

Change other tests so that we remove cost figures or make the preferred
plan less sensitive to changes in the cost of hash joins.

We do this before implementing the new cost function, so that it is easier
to see what the effect of the new cost function is.

Change-Id: I06f14bafd96f1001c565a7e601847a0b1bac6703