Bug#37733272: Planning overhead for ordered range queries in hypergraph [1, FDs, noclose]
If there is a unique index on a column col0 in a table, the hypergraph
optimizer collects functional dependencies col0->col1, col0->col2,
..., col0->colN for all the other columns in the table.

Many of these functional dependencies might not be useful for
optimizing the query. Such dependencies are often pruned away later,
in LogicalOrderings::PruneFDs(), but the cost of adding them and
pruning them can be high, so it's better not to add them in the first
place.

This patch reduces the number of functional dependencies collected
from unique indexes by not adding dependencies on the columns that are
not referenced in the query.

It also ignores extended key parts when analyzing unique indexes. When
including extended key parts, even non-unique secondary indexes are
considered unique. The only reason they are considered unique, is that
they include the primary key columns. Since we already get functional
dependencies reflecting that the primary key is unique when we collect
functional dependencies from the primary key, we don't need to spend
time adding more functional dependencies reflecting the same facts
from the secondary indexes.

Change-Id: I727296498cc7ffa3bb3da0ae1d62f98d73245941