WL#16570: Migrate the firewall plugin to a component
MySQL Enterprise Firewall is an application-level firewall that enables
database administrators to permit or deny SQL statement execution based
on matching against lists of accepted statement patterns.

This WL implements the firewall as a component and deprecates the firewall
plugin. The functionality of the component follows functionality of
the plugin except deprecated features. WL provides also tools for
migrating from firewall plugin to component and vice versa.

Change-Id: I1a381b3c33182c5ced0957373a9655028d8472a8