Bug#37742092 Limit number of spill sets in set operations spill-to-disk handling
Cf. also abandoned patch for Bug#35663846 WL15730 Unable to abort
select into query executed on lakehouse tables

We saw an assert in INTERSECT because the computed number of chunks
exceed 65535, see assert in hash_to_chunk_index.  This can happen if
the number of estimated rows is very high (in the repro case it is way
too high), or because the setting of set_operations_buffer_size is too
low.

There is a missing check for this situation in
materialize_iterator::SpillState::compute_chunk_file_sets.  This patch
adds a check.  If we need more than 2 (SpillState::m_max_sets) chunk
file sets, we revert to index tmp table based de-duplication. To alert
the user, this information is presented in

  - a note for the query, and in
  - optimizer trace

The note contains hint to check optimizer trace for recommendations.

Change-Id: Ib8fa49c94116a91b79c4349221a027cb014623ca