Bug#37733272: Planning overhead for ordered range queries in hypergraph [2, GetHandle, noclose]
When the hypergraph optimizer creates orderings and functional
dependencies from indexes, it generates Item_field objects for each
column in the index, and possibly many times per column.

The size of the Item_field objects is non-trivial, so allocating them
unnecessarily costs time and memory. They are created because
LogicalOrderings::GetHandle() needs an Item in order to retrieve an
Item handle. If LogicalOrderings already has an Item for the column,
which it often has, the newly created Item_field is not needed after
the handle has been returned by GetHandle().

To reduce the amount of Item_fields created by the hypergraph
optimizer, LogicalOrderings::GetHandle() is overloaded with a variant
that takes a Field instead of an Item. If an Item_field for the same
Field already exists in LogicalOrderings, a handle to the existing
Item_field is returned. Only if one is not found, a new Item_field is
allocated and added to LogicalOrderings::m_items.

Change-Id: Ia074c73d37aee03a57e783cc59113f32a91ad83a