Bug#37733272: Planning overhead for ordered range queries in hypergraph [3, homogenize, noclose]
The hypergraph optimizer's interesting orders framework performs an
operation called homogenization which aims to reduce the number of
tables referenced by an ordering to one, so that the ordering can be
pushed down as a sort-ahead on top of a single table.

In single-table queries, this is unnecessary, as no ordering could
refer to more than one table in such a query. Profiling suggests it
takes a noticeable chunk of time for some single-table ordered
queries. This patch skips the homogenization step if the collected
interesting orders and functional dependencies refer to a single
table.

Change-Id: I0709c467ff2db102b83a4f388bfcacdd69c1dec5