Bug#37748564 MRS passthroughDbUser must not keep user passwords in memory
Post push fix for WL#16741

The MRS passthroughDbUser option enables REST services to use the
requesting users MySQL account to execute MySQL requests when serving
REST requests. The implementation stored the password used during MRS
authentication of MySQL Internal type users in memory, associated to
the HTTP session. Future requests from that HTTP session would then
open MySQL connections with the saved credentials.

This implementation is now changed to no longer keep passwords beyond
the scoped of the HTTP authentication request.

When a MySQL Internal account is authenticated for a service with
passthroughDbUser enabled, a connection pool is now created and
pre-populated for that HTTP session. Future requests from that
session will be served using connections from that pool. If the
number of concurrent requests exceeds the size of the connection pool,
they will be rejected with status 429 TooManyRequests.

The connection pool is destroyed when the HTTP session is closed,
either by expiring, timing out from idleness or explicit logout.

Change-Id: If96755dad916a2c178b5299cdf187d70d86b675e