Bug#37733272: Planning overhead for ordered range queries in hypergraph [4, bytes_per_row, noclose]
The hypergraph optimizer estimates the number of bytes it has to read
per row of a table, and uses this estimate to calculate the cost of
index lookups, table scans and index range scans. The estimate is
recalculated for each access method, even though the estimate is
independent of the access method and could be reused.

Profiling suggests that a noticeable amount of time is spent on
calculating this estimate in single-table queries that have multiple
possibilities for index access methods. It is especially noticeable
when the row size of the table is so large that
EstimateBytesPerRowTable() calls EstimateBytesPerRowWideTable() to get
a more sophisticated estimate.

To reduce the overhead, this patch caches the estimate in the TABLE
object so that it is calculated only once per table.

Change-Id: Ic8cbcd0f230cd3df0da74ff9ce426588ecd2de1a