Bug#37758147: The test rpl.rpl_semi_sync_binlog_suffix_exceed_six_digit fails on pb2 daily trunk
Problem:

The test rpl_semi_sync_binlog_suffix_exceed_six_digit.test which was added
as part of BUG#37024069 is failing when run with gtid_mode=ON

Analysis:

The statement CHANGE REPLICATION SOURCE SOURCE_LOG_FILE gives error when
running with gtid_mode=ON as SOURCE_LOG_FILE cannot be set when
SOURCE_AUTO_POSITION is active.

Fix:

We are not adding any extra coverage by running this test with gtid_mode=ON
so move the test from rpl suite to rpl_nogtid.

Change-Id: I86da78a1d5406b0e497abe8fbb0b2f596f31c85f