Bug#37765153	Allow debian/ubuntu builds to run rootless
In the deb packaging we run a few install commands that change ownership
of files to the root user. This means the build has to run as root,
or with fakeroot.

This isn't recommended or required (downstream packaging installs these
files without changing ownership), and in Ubuntu 25.04 the default config
does not allow builds to use root, so it fails.

Removed the ownership change from these calls to let the build run rootless.

Change-Id: I318693eb11ffc193c2d8b43c746685786f8366e7