Bug#34479495: Hypergraph optimizer: Multicolumn index statistics is not used
In some cases, the hypergraph optimizer fails to use multi-column
index statistics for estimating the selectivity of join conditions.
This happens if the join condition on one of the indexed columns is
cannot be pushed down to a join in PushDownCondition(), but is instead
left for creating a cycle edge in the join graph later.

For example, in TPC-DS Q64, this happens with the join condition
between store_sales and store_returns. In the optimizer trace, we see
that the join condition between those tables leads to the creation of
a cycle edge:

    Pushing conditions down.
    - condition (store_sales.ss_ticket_number = store_returns.sr_ticket_number) induces a hypergraph cycle

And then, for the calculation of the selectivity of the full join
condition between those tables, the trace says this:

    Selectivity of join (store_sales.ss_ticket_number = store_returns.sr_ticket_number) AND (store_sales.ss_item_sk = store_returns.sr_item_sk):
     - estimating selectivity 4.16667e-06 for field store_sales.ss_ticket_number from histogram showing 240000 distinct values and non-null fraction 1.
     - estimating selectivity 5.89372e-06 for field store_returns.sr_ticket_number from histogram showing 169672 distinct values and non-null fraction 1.
     - Using histogram statistics:  selectivity = 0.000006
     - selectivity for multiple equal(store_sales.ss_ticket_number, store_returns.sr_ticket_number) = 5.89372e-06
     - found 1-field prefix of candidate index PRIMARY with selectivity 5.55556e-05 for last field store_sales.ss_item_sk
     - capping selectivity to 5.55556e-05 since index is unique
     - found 1-field prefix of candidate index PRIMARY with selectivity 5.55679e-05 for last field store_returns.sr_item_sk
     - Using index statistics: selectivity = 0.000056
     - selectivity for multiple equal(store_sales.ss_item_sk, item.i_item_sk, store_returns.sr_item_sk, cs_ui.cs_item_sk) = 5.55556e-05

That is, even though both store_sales and store_returns have a primary
key on (sr_item_sk, sr_ticket_number), the optimizer uses the index
statistics on the prefix column (sr_item_sk) only. For the other
column in the join condition (ss_ticket_number), the selectivity is
calculated separately using histograms, and under the assumption that
the two predicates are independent.

If instead the selectivity of the two join predicates is based on the
multi-column statistics of the primary keys in the two tables, the
selectivity estimate also accounts for the correlation between the two
columns, giving a better estimate.

This patch makes AddCycleEdges() record the equijoin predicates of
cycle edges in the CompanionSet in the same way as it's done for
equijoin predicates in the other edges. This makes the hypergraph
optimizer able to use multi-column index statistics also for these
join conditions. After the patch, the optimizer trace shows that
multi-column statistics are indeed used:

    Selectivity of join (store_sales.ss_ticket_number = store_returns.sr_ticket_number) AND (store_sales.ss_item_sk = store_returns.sr_item_sk):
     - found 2-field prefix of candidate index PRIMARY with selectivity 0.00624912 for last field store_sales.ss_ticket_number
     - found 2-field prefix of candidate index PRIMARY with selectivity 0.0625917 for last field store_returns.sr_ticket_number
     - Using index statistics: selectivity = 0.062592
     - selectivity for multiple equal(store_sales.ss_ticket_number, store_returns.sr_ticket_number) = 0.0625917
     - found 1-field prefix of candidate index PRIMARY with selectivity 5.55556e-05 for last field store_sales.ss_item_sk
     - capping selectivity to 5.55556e-05 since index is unique
     - found 1-field prefix of candidate index PRIMARY with selectivity 5.55679e-05 for last field store_returns.sr_item_sk
     - Using index statistics: selectivity = 0.000056
     - selectivity for multiple equal(store_sales.ss_item_sk, item.i_item_sk, store_returns.sr_item_sk, cs_ui.cs_item_sk) = 5.55556e-05

Change-Id: I91ca775465b0a575872d5adb3b3c82de1571acc8